feat: Add SingleVote component
