+++ apps/web/src/views/Voting/Proposal/Vote.tsx
@@ -11,7 +11,6 @@ import {
   Image,
   Message,
   MessageText,
-  Radio,
   Text,
   useModal,
   useToast,
@@ -21,8 +20,9 @@ import { getBalanceNumber } from '@pancakeswap/utils/formatBalance'
 import ConnectWalletButton from 'components/ConnectWalletButton'
 import { useVeCakeBalance } from 'hooks/useTokenBalance'
 import { useState } from 'react'
-import { Proposal, ProposalState } from 'state/types'
-import { styled } from 'styled-components'
+import { Proposal, ProposalState, ProposalTypeName } from 'state/types'
+import { SingleVote } from 'views/Voting/Proposal/VoteType/SingleVote'
+import { State } from 'views/Voting/Proposal/VoteType/types'
 import { useAccount } from 'wagmi'
 import CastVoteModal from '../components/CastVoteModal'
 
@@ -31,30 +31,6 @@ interface VoteProps extends CardProps {
   hasAccountVoted: boolean
   onSuccess?: () => void
 }
-
-interface State {
-  label: string
-  value: number
-}
-
-const Choice = styled.label<{ isChecked: boolean; isDisabled: boolean }>`
-  align-items: center;
-  border: 1px solid ${({ theme, isChecked }) => theme.colors[isChecked ? 'success' : 'cardBorder']};
-  border-radius: 16px;
-  cursor: ${({ isDisabled }) => (isDisabled ? 'not-allowed' : 'pointer')};
-  display: flex;
-  margin-bottom: 16px;
-  padding: 16px;
-`
-
-const ChoiceText = styled.div`
-  flex: 1;
-  padding-left: 16px;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  width: 0;
-`
 
 const Vote: React.FC<React.PropsWithChildren<VoteProps>> = ({ proposal, hasAccountVoted, onSuccess, ...props }) => {
   const [vote, setVote] = useState<State>({
@@ -103,35 +79,9 @@ const Vote: React.FC<React.PropsWithChildren<VoteProps>> = ({ proposal, hasAccou
         </Flex>
       </CardHeader>
       <CardBody>
-        {proposal.choices.map((choice, index) => {
-          const isChecked = index + 1 === vote.value
-
-          const handleChange = () => {
-            setVote({
-              label: choice,
-              value: index + 1,
-            })
-          }
-
-          return (
-            <Choice key={choice} isChecked={isChecked} isDisabled={!account}>
-              <div style={{ flexShrink: 0 }}>
-                <Radio
-                  scale="sm"
-                  value={choice}
-                  checked={isChecked}
-                  onChange={handleChange}
-                  disabled={!account || proposal.state === ProposalState.CLOSED}
-                />
-              </div>
-              <ChoiceText>
-                <Text as="span" title={choice}>
-                  {choice}
-                </Text>
-              </ChoiceText>
-            </Choice>
-          )
-        })}
+        {proposal.type === ProposalTypeName.SINGLE_CHOICE && (
+          <SingleVote proposal={proposal} vote={vote} setVote={setVote} />
+        )}
         {account ? (
           <>
             {hasAccountVoted ? (
