+++ apps/web/src/views/Voting/Proposal/VoteType/types.ts
@@ -1,0 +1,4 @@
+export interface State {
+  label: string
+  value: number
+}
